app/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
psycopg
psycopg-binary
pydantic_settings
alembic
pydantic
aiogram
python-jose
redis==5.0.4
cryptography==42.0.5 
Jinja2==3.1.3
passlib
python-multipart
pydantic[email]